- making gcc happier (silencing comparison warnings and unused vars)

+++ sensitivity/sensitiveDetector.cc
@@ -79,7 +79,7 @@ vector<BackgroundHit*> sensitiveDetector::getNextBackgroundEvent()
 	if(backgroundHits != nullptr) {
 
 		backgroundEventNumber++;
-		if(backgroundEventNumber == backgroundHits->size()) {
+		if(backgroundEventNumber == (int) backgroundHits->size()) {
 			backgroundEventNumber = 1;
 		}
